Fix quality page indicators page
