+++ decidim-participatory_processes/app/commands/decidim/participatory_processes/create_space_page.rb
@@ -1,0 +1,41 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ParticipatoryProcesses
+    class CreateSpacePage < Decidim::Command
+      # Public: Initializes the command.
+      #
+      # organization - A object with the params.
+      def initialize(organization)
+        @organization = organization
+      end
+
+      # Executes the command.
+      # i18n-tasks-use t('decidim.system.default_pages.terms-of-service')
+      #
+      # Returns nothing.
+      def call
+        StaticPage.find_or_create_by!(organization:, slug: "democratic-quality-indicators") do |page|
+          page.decidim_organization_id = organization.id
+          page.title = localized_attribute(organization, :title)
+          page.content = localized_attribute(organization, :content)
+          page.allow_public_access = true
+        end
+      end
+
+      private
+
+      attr_reader :organization
+
+      def localized_attribute(organization, attribute)
+        organization.available_locales.inject({}) do |result, locale|
+          text = I18n.with_locale(locale) do
+            I18n.t(attribute, scope: "decidim.participatory_processes.static_pages.democratic_quality_indicators")
+          end
+
+          result.update(locale => text)
+        end
+      end
+    end
+  end
+end
