+++ decidim-participatory_processes/lib/decidim/participatory_processes/seeds.rb
@@ -7,6 +7,8 @@ module Decidim
   module ParticipatoryProcesses
     class Seeds < Decidim::Seeds
       def call
+        Decidim::ParticipatoryProcesses::CreateSpacePage.call(organization)
+
         create_content_block!
 
         process_groups = []
